# server/public/openapi.yaml
openapi: 3.1.0
info:
  title: TeamSpark Pachinko API
  version: 1.0.0
  description: |
    HTTP API for policy configuration, message and alert review, analytics,
    Arcade Engine webhooks, retention, and application settings.
servers:
  - url: http://localhost:3000/api/v1
    description: API base (Next.js app)
...
components:
  schemas:
    Meta:
      type: object
      properties:
        apiVersion:
          type: string
          enum: ["1.0"]
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Status message
...
    AggregateResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
                description: The value of the dimension
              count:
                type: integer
                description: The count of messages for this value
              success:
                type: integer
                description: The count of successful messages
              error:
                type: integer
                description: The count of error messages
        query:
          type: object
          properties:
            dimension:
              type: string
              description: The dimension being queried
            timeRange:
              type: object
              properties:
                start:
                  type: string
                  format: date-time
                  description: Start of the time range
                end:
                  type: string
                  format: date-time
                  description: End of the time range
            filters:
              type: object
              properties:
                serverName:
                  type: string
                userId:
                  type: string
                clientId:
                  type: string
                method:
                  type: string
                sourceIP:
                  type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              meta:
                $ref: '#/components/schemas/Meta'
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            type: object
            properties:
              meta:
                $ref: '#/components/schemas/Meta'
paths:
...
  /analytics/messages/timeSeries:
    get:
      tags: [Analytics]
      summary: Message time series
      description: |
        The dimension query parameter accepts payloadMethod, payloadToolName, userId, source,
        payloadToolkit, or aliases method, toolName, serverName (mapped to payload fields).
      parameters:
        - name: dimension
          in: query
          required: true
          schema:
            type: string
            enum:
              - payloadMethod
              - payloadToolName
              - userId
              - source
              - payloadToolkit
              - method
              - toolName
              - serverName
        - name: timeUnit
          in: query
          required: true
          schema: { type: string, enum: [hour, day, week, month] }
        - name: userId
          in: query
          schema: { type: string }
        - name: source
          in: query
          schema: { type: string }
        - name: payloadToolkit
          in: query
          schema: { type: string }
        - name: payloadMethod
          in: query
          schema: { type: string }
        - name: payloadToolName
          in: query
          schema: { type: string }
        - name: startTime
          in: query
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: timeSeries payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: "#/components/schemas/Meta"
                  timeSeries:
                    type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /analytics/messages/aggregate:
    get:
      tags: [Analytics]
      summary: Message aggregates
      description: Same dimension names and aliases as time series; optional startTime and endTime filters.
      parameters:
        - name: dimension
          in: query
          required: true
          schema: { type: string }
        - name: userId
          in: query
          schema: { type: string }
        - name: source
          in: query
          schema: { type: string }
        - name: payloadToolkit
          in: query
          schema: { type: string }
        - name: payloadMethod
          in: query
          schema: { type: string }
        - name: payloadToolName
          in: query
          schema: { type: string }
        - name: startTime
          in: query
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: aggregate payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /analytics/dimensions:
    get:
      tags: [Analytics]
      summary: Dimension values for filters
      description: |
        Repeat the dimension query parameter for each axis. Supported values are userId, source,
        payloadToolkit, payloadMethod, payloadToolName, policyId, conditionName, seen, and severity.
      parameters:
        - name: dimension
          in: query
          required: true
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              enum:
                - userId
                - source
                - payloadToolkit
                - payloadMethod
                - payloadToolName
                - policyId
                - conditionName
                - seen
                - severity
        - name: userId
          in: query
          schema: { type: string }
        - name: source
          in: query
          schema: { type: string }
        - name: payloadToolkit
          in: query
          schema: { type: string }
        - name: payloadMethod
          in: query
          schema: { type: string }
        - name: payloadToolName
          in: query
          schema: { type: string }
        - name: policyId
          in: query
          schema: { type: integer }
        - name: conditionName
          in: query
          schema: { type: string }
        - name: seen
          in: query
          schema: { type: boolean }
        - name: severity
          in: query
          schema: { type: integer }
        - name: startTime
          in: query
          schema: { type: string, format: date-time }
        - name: endTime
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...